add code to ext + dependencies
