+++ .gitignore
@@ -1,3 +1,4 @@
 /Manifest.toml
 /.vscode
 .DS_Store
+/meus_testes
